# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/maximumspecification": {
            "get": {
                "description": "Get a list of MaximumSpecification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaximumSpecification"
                ],
                "summary": "Get all MaximumSpecification.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MaximumSpecification"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new MaximumSpecification.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaximumSpecification"
                ],
                "summary": "Create New MaximumSpecification.",
                "parameters": [
                    {
                        "description": "the body to create a new MaximumSpecification",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.maximumSpecificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaximumSpecification"
                        }
                    }
                }
            }
        },
        "/maximumspecification/{id}": {
            "get": {
                "description": "Get an MaximumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaximumSpecification"
                ],
                "summary": "Get MaximumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MaximumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaximumSpecification"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a MaximumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaximumSpecification"
                ],
                "summary": "Delete one MaximumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MaximumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update MaximumSpecification by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MaximumSpecification"
                ],
                "summary": "Update MaximumSpecification.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MaximumSpecification id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update MaximumSpecification",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.maximumSpecificationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MaximumSpecification"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.maximumSpecificationInput": {
            "type": "object",
            "properties": {
                "directx": {
                    "type": "string"
                },
                "game_id": {
                    "type": "integer"
                },
                "graphics": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "processor": {
                    "type": "string"
                },
                "soundcard": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                }
            }
        },
...
        "models.MaximumSpecification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "directx": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "graphics": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "memory": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "processor": {
                    "type": "string"
                },
                "soundcard": {
                    "type": "string"
                },
                "storage": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}